Fix compilation on 64-bit machines.
On 64-bit machines the construction of QVariant from intptr_t is
ambigious as it could be void* or ulonglong. The documentation for
the debugger extension on the other hand is clear that it should be a
qint64. An explicit construction of a qint64 fixes it.

+++ src/script/api/qscriptengineagent_p.h
@@ -114,7 +114,7 @@ public:
     virtual void didReachBreakpoint(const JSC::DebuggerCallFrame&, intptr_t sourceID, int lineno, int column)
     {
         QList<QVariant> args;
-        args << sourceID << lineno << column;
+        args << qint64(sourceID) << lineno << column;
         if (q_ptr->supportsExtension(QScriptEngineAgent::DebuggerInvocationRequest))
             q_ptr->extension(QScriptEngineAgent::DebuggerInvocationRequest, args);
     };
